ISC1
Add second route prototype to pilot
Note that this is not fully usable - the end includes many repeated frames that are just there to close the gap between start and end with 360 frames. The original video is only 55 seconds - 5 seconds short of a complete loop. Just putting this in to test.